+++ package.json
@@ -22,6 +22,7 @@
   },
   "homepage": "https://github.com/INFURA/documentation#readme",
   "dependencies": {
+    "dotenv": "^6.0.0",
     "fs": "0.0.1-security",
     "git-origin-url": "^0.3.0",
     "mkdirp": "^0.5.1",
@@ -29,6 +30,8 @@
     "yamljs": "^0.3.0"
   },
   "devDependencies": {
-    "dotenv": "^6.0.0"
+    "eslint": "^5.4.0",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-plugin-import": "^2.14.0"
   }
 }
